Add server function and refactor
