+++ server.js
@@ -3,6 +3,7 @@
 const hof = require('hof');
 const config = require('./config');
 const mockPostcode = require('./mock-postcode');
+const _ = require('lodash');
 
 let settings = require('./hof.settings');
 
@@ -24,5 +25,25 @@ app.use((req, res, next) => {
   res.locals.htmlLang = 'en';
   next();
 });
+
+if (config.env === 'development' || config.env === 'test') {
+  app.use('/test/bootstrap-session', (req, res) => {
+    const appName = req.body.appName;
+
+    if (!_.get(req, 'session[`hof-wizard-${appName}`]')) {
+      if (!req.session) {
+        throw new Error('Redis is not running!');
+      }
+
+      req.session[`hof-wizard-${appName}`] = {};
+    }
+
+    Object.keys(req.body.sessionProperties || {}).forEach(key => {
+      req.session[`hof-wizard-${appName}`][key] = req.body.sessionProperties[key];
+    });
+
+    res.send('Session populate complete');
+  });
+}
 
 module.exports = app;
