+++ test/helpers/supertest_session/session-data/gro/pages/additional.js
@@ -1,0 +1,5 @@
+
+module.exports = {
+  'additional-text': 'n/a',
+  'additional-radio': 'no'
+};
