+++ test/helpers/supertest_session/session-data/gro/pages/address.js
@@ -1,0 +1,7 @@
+
+module.exports = {
+  building: '10 Downing Street',
+  townOrCity: 'London',
+  countyOrState: 'Greater London',
+  postcodeOrZIPcode: 'W12 8DE'
+};
