+++ test/helpers/supertest_session/session-data/gro/pages/confirm.js
@@ -1,0 +1,2 @@
+
+module.exports = {};
