+++ test/helpers/supertest_session/session-data/gro/pages/details.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+  'details-text': 'n/a',
+  'existing-radio': 'yes',
+  'previous-radio': 'yes'
+};
