+++ test/helpers/supertest_session/session-data/gro/pages/email-address.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+  'email-text': 'email@test.com'
+};
