+++ test/helpers/supertest_session/session-data/gro/pages/how.js
@@ -1,0 +1,5 @@
+
+module.exports = {
+  'how-radio': 'online',
+  'online-toggle-text': 'COL12345/2021'
+};
