+++ test/helpers/supertest_session/session-data/gro/pages/people.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+  'person-one': 'Ben Smith'
+};
