+++ test/helpers/supertest_session/session-data/gro/pages/person.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+  'person-text': 'Ben Smith'
+};
