+++ test/helpers/supertest_session/session-data/gro/pages/type.js
@@ -1,4 +1,4 @@
 
 module.exports = {
-
+  'type-radio': 'birth'
 };
