+++ test/helpers/supertest_session/session-data/gro/pages/when.js
@@ -1,0 +1,7 @@
+
+module.exports = {
+  'when-date-day': '12',
+  'when-date-month': '12',
+  'when-date-year': '2021',
+  'when-date': '2021-12-12'
+};
