+++ test/helpers/supertest_session/session-data/gro/pages/which.js
@@ -1,0 +1,4 @@
+
+module.exports = {
+  'which-radio': 'standard'
+};
